Change implementation of |BitSet<n>| to use chuncks of 32 bits, not maybe 64
The chunck size was |unsigned long| which is 64 bits on some systems. This is
unnecessarily large for the chunk size in a |BitSet|. The result was that even
a |BitSet| holding just a few bits could claim 64 bits space in a structure,
or even more in order to assure alignment on a 8-byte boundary. As all
template instances are explicit, there need be no penalty for a small chunk
size; with the current setting of |RANK_MAX=32| there is however no need for
very small chunks either, so 32 bits (type |uint_least32_t|) is OK for now.

We needed to move one constructor from |BitSetBase<1>| to |BitSetBase<2>| a
the smaller chunk size together with |RANK_MAX==32| made this necessary.

In passing, we changed the template parameter from |size_t| to |unsigned int|.
Since the parameter designates the number of bits in what is basically a wide
unsigned integral type, the range suggested by |size_t| is way exaggerated.

+++ sources/utilities/bitset_fwd.h
@@ -24,7 +24,7 @@ namespace atlas {
 
 namespace bitset {
 
-  template<size_t n> class BitSet;
+  template<unsigned int n> class BitSet;
 
 }
